+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Change Log -- ietfdata
  - Add `Datatracker::meeting()` method
  - Update `Datatracker::meetings()` to take `start_date` and `end_date`
    parameters rather than `since` and `until`.
- - Add `Datatracker::schedule()` method and `Schedule` type
+ - Add `Datatracker::meeting_schedule()` method 
+ - Add `Datatracker::Schedule` type
  - Add `Datatracker::Timeslot` type
  - Add `Datatracker::Assignment` type